remove builder & refactor log path
