+++ Engine/Source/logging/src/layers/file.rs
@@ -1,7 +1,6 @@
 use std::{
     fs::{File, OpenOptions},
     io::Write,
-    path::Path,
     sync::Mutex,
 };
 
@@ -12,9 +11,8 @@ use super::format::{
     extract_event_data, format_level, format_line, format_timestamp, format_timestamp_filename,
 };
 
-// ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-// FileLayer
-// ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+// TODO: add compile time var with proper path
+const LOG_PATH: &str = "LOG_PATH";
 
 /// A [`tracing_subscriber::Layer`] that writes **plain-text** (no ANSI codes)
 /// log lines to two files simultaneously:
@@ -40,9 +38,8 @@ impl FileLayer {
     ///
     /// Returns an error if the directory cannot be created or either file
     /// cannot be opened.
-    pub fn new(log_dir: impl AsRef<Path>) -> std::io::Result<Self> {
-        let dir = log_dir.as_ref();
-        std::fs::create_dir_all(dir)?;
+    pub fn new() -> std::io::Result<Self> {
+        std::fs::create_dir_all(LOG_PATH)?;
 
         let ts = format_timestamp_filename(&time::OffsetDateTime::now_utc());
 
@@ -51,14 +48,14 @@ impl FileLayer {
         let timestamped = OpenOptions::new()
             .create(true)
             .append(true)
-            .open(dir.join(format!("{ts}.log")))?;
+            .open(format!("{LOG_PATH}/{ts}.log"))?;
 
         // Truncate latest.log so it only contains the current session.
         let latest = OpenOptions::new()
             .create(true)
             .write(true)
             .truncate(true)
-            .open(dir.join("latest.log"))?;
+            .open(format!("{LOG_PATH}/latest.log"))?;
 
         Ok(Self {
             writers: Mutex::new(FileWriters {
